--- a/assets/syntax.txt
+++ b/assets/syntax.txt
@@ -1,4 +1,4 @@
-_seps: "(){}[],.:;=<>*·+-/%^?|&∧∨!¬∑∃∀"
+_seps: "(){}[],.:;=<>*·+-/%^?|&∧∨!¬∑∃∀\n"
 
 200 multi_line_comment = ["/*" ..."*/"? .r?({
     [!"*/" "*" ..."*/"?] [multi_line_comment ..."*/"?] ["/" ..."*/"?]
@@ -107,7 +107,13 @@ _seps: "(){}[],.:;=<>*·+-/%^?|&∧∨!¬∑∃∀"
 
 101 + = [?w {"+":"+" "||":"+" "∨":"+" "or":"+"} ?w]
 102 - = [?w "-":"-" ?w]
-103 * = [?w {"*.":"*." "·":"*." "*":"*" "&&":"*" "∧":"*" "and":"*"} ?w]
+// Allow whitespace before multiplication sign, but no new line.
+// This prevents `x` on a new line from being interpreted as multiplication sign.
+103 * = [.r?({" " "\t" "\r"}) {
+    "*.":"*." "·":"*."
+    "x":"x" "⨯":"x"
+    "*":"*" "&&":"*" "∧":"*" "and":"*"
+} ?w]
 104 / = [?w "/":"/" ?w]
 105 % = [?w "%":"%" ?w]
 107 pow = [lexpr:"base" ?w "^" ?w lexpr:"exp"]