--- a/source/syntax/vec4.rs
+++ b/source/syntax/vec4.rs
@@ -1,0 +1,13 @@
+fn main() {
+    a := (0, 0, 1)
+    b := (1, 0, 0)
+    c := a ⨯ b
+    d := a · b
+    println(c)
+    println(d)
+
+    c := a x b
+    d := a *. b
+    println(c)
+    println(d)
+}