--- a/source/test.rs
+++ b/source/test.rs
@@ -1,6 +1,6 @@
 fn main() {
-    a := (1, 2, 3, 4)
-    b := 3
-    c := 3 / a
-    println(c)
+    v := (1, 0, 0)
+    u := (0, 1, 0)
+    x := v x u
+    println(x)
 }