See https://github.com/PistonDevelopers/dyon/issues/130

Design: https://github.com/PistonDevelopers/dyon/issues/144

### Performance

This leads to 2.8x improvement in performance for the n_body example:

```
Before:
test tests::bench_add         ... bench:  43,393,867 ns/iter (+/- 6,806,476)
test tests::bench_add_n       ... bench:  15,092,144 ns/iter (+/- 3,001,957)
test tests::bench_array       ... bench:  18,773,838 ns/iter (+/- 3,739,800)
test tests::bench_call        ... bench:  17,788,675 ns/iter (+/- 3,720,407)
test tests::bench_len         ... bench:   3,382,249 ns/iter (+/- 205,532)
test tests::bench_main        ... bench:   1,891,454 ns/iter (+/- 157,257)
test tests::bench_min         ... bench:  16,661,012 ns/iter (+/- 2,141,291)
test tests::bench_min_fn      ... bench:  81,098,014 ns/iter (+/- 7,106,106)
test tests::bench_n_body      ... bench: 208,976,966 ns/iter (+/- 18,541,989) <---
test tests::bench_object      ... bench:  27,482,168 ns/iter (+/- 5,447,122)
test tests::bench_primes      ... bench:  48,065,297 ns/iter (+/- 5,965,289)
test tests::bench_primes_trad ... bench:  45,867,012 ns/iter (+/- 6,662,685)
test tests::bench_sum         ... bench:   6,830,854 ns/iter (+/- 614,111)

After:
test tests::bench_add         ... bench:  42,161,587 ns/iter (+/- 5,777,769)
test tests::bench_add_n       ... bench:  15,811,724 ns/iter (+/- 3,484,295)
test tests::bench_array       ... bench:  18,372,826 ns/iter (+/- 3,036,913)
test tests::bench_call        ... bench:  17,736,857 ns/iter (+/- 2,506,030)
test tests::bench_len         ... bench:   3,516,452 ns/iter (+/- 187,508)
test tests::bench_main        ... bench:   2,004,404 ns/iter (+/- 541,539)
test tests::bench_min         ... bench:  17,647,011 ns/iter (+/- 3,824,137)
test tests::bench_min_fn      ... bench:  81,071,483 ns/iter (+/- 7,499,528)
test tests::bench_n_body      ... bench:  75,921,824 ns/iter (+/- 6,012,761) <---
test tests::bench_object      ... bench:  26,914,886 ns/iter (+/- 5,105,781)
test tests::bench_primes      ... bench:  47,865,836 ns/iter (+/- 4,815,001)
test tests::bench_primes_trad ... bench:  45,441,088 ns/iter (+/- 5,097,709)
test tests::bench_sum         ... bench:   6,895,172 ns/iter (+/- 333,311)

(old_bench_n_body - old_bench_main) / (new_bench_n_body - new_bench_main)
= (208976966 - 1891454) / (75921824 - 2004404)
= 2.8015792759000515
```